requirements.txt
pyserial>=3.5
dbus-python>=1.3.2
